HOCS-1674 Add internal ingress
This commit adds an internal ingress to supplement the external
ingresses that already exist, and makes the cs-dev and cs-qa environments
use the internal ingress instead of the external one. You can choose either
ingress for any given environment, by modifying its entry in the
deploy.sh script -- although you cannot have both.

Currently, you can only use an external ingress if your IP address is
in the allowed list of IP addresses. This, in practice, is office
locations and does not include devices working elsewhere, such as from
home.

This meant that people wishing to access the application from a
non-standard location needed to rely on workarounds. The ACP "Tunnel
All" VPN is often used for this purpose, which routes all traffic from a
device over an ACP AWS VPC (that is, our hosting provider's network),
and has a side-effect of changing the apparent IP address to one
permissible by the IP allowlist. This wasn't efficient -- especially
apparent when one tries to log into the system whilst simultaneously
using video conferencing.

We can fix this by exposing an internal VPN-accessible ingress as an
optional replacement for our our existing Internet-addressable external
ingress, making it much easier for developers to access the service from
any location, with the downside of making it more complex for "normal"
users on corporate hardware.

Environments using the internal ingress instead of the external one will
no longer be accessible via the Internet, and therefore will not be
available via Poise by default. Anybody wanting to connect must access
the service via the "kube-platform" ACP VPN. Most users of the
non-production environments are connecting to the kube-platform VPN
routinely anyway: most of the development toolset, such as Drone CI's
web frontend and log aggregation frontend Kibana, are already using
internal ingresses.

Kubernetes allows you to have lots of ingresses, but our app doesn't
support it. Specifically, the redirection URL setting of Keycloak
Gatekeeper can only be set once, so although you can authenticate from
an arbitrary hostname, it is difficult (or at least, annoying) to return
to the app after being dealt with by the central Keycloak server.

Theoretically we could spin up a second hocs-frontend pod with a
different redirection URL set, but at the moment that introduces more
complexity than an additional ingress is worth.

We decide which ingress to deploy based on the hostname: if the DNS prefix
contains "internal", we use the internal configuration instead of the
external one.

The two ingress configurations are broadly similar. The main difference,
other than the labels used to tell ACP which type we want, is the way we
obtain certificates from Let's Encrypt. Normally Let's Encrypt (our TLS
certificate provider) uses a HTTP request to validate that the person
requesting a certificate actually controls the associated hostname.
That's not possible with an internal ingress, as it's not routed over
the Internet. Instead, we tell it to confirm via the Domain Name System.

This commit only changes the behaviour for the cs-dev and cs-qa
environments now. Subsequent environments will be introduced in
subsequent commits.

For this change to work an updated network policy had to be updated so
the management UI and the new internal ingress to communicate with each
other. Network policies are managed centrally in a different repo, at
UKHomeOffice/kube-hocs, and this change was made for the equivalent
hocs-frontend ingress addition.

See also: https://github.com/UKHomeOffice/kube-hocs-frontend/pull/73