HOCS-3061 Use multiple ingresses
This commit removes the hard-coded redirection URL from the
keycloak-gatekeeper (keycloak-proxy) configurations for non-production
environments. This means keycloak-gatekeeper will try and guess the
correct values from the host header. That way we can use multiple
ingresses.

This makes it slightly less secure, so we don't do it for production. As
we don't do it for production, we also don't do it for demo -- so we
have at least one non-production environment in our pipeline with
similar settings to notice any (unlikely) problems.

Due to limitations with how the host header works, this commit breaks
the logout link on most environments, as Keycloak tries to redirect to
http://{endpoint} not https://{endpoint} which we've disallowed in our
Keycloak realm configuration.

This is the MUI equivalent of the frontend-change
81bb909d1eae819161f64f58c333256ca529da03.